// apps/web/src/prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] }
// ...
export type ArticleDocumentDataBodySlice =
  | RichTextSlice
  | VideoEmbedSlice
  | DrillDiagramSlice
  | CalloutBoxSlice
  | AdPlacementSlice
  | PrintableDownloadSlice

export interface ArticleDocumentData {
  title: prismic.KeyTextField
  sport: prismic.ContentRelationshipField<'sport'>
  age_group: prismic.ContentRelationshipField<'age_group'>
  tags: prismic.KeyTextField
  featured_image: prismic.ImageField
  summary: prismic.KeyTextField
  body: prismic.SliceZone<ArticleDocumentDataBodySlice>
  printable_pdf: prismic.LinkField
  published_date: prismic.DateField
  seo_title: prismic.KeyTextField
  seo_description: prismic.KeyTextField
}

export type ArticleDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<Simplify<ArticleDocumentData>, 'article', Lang>